refactor: Clarify Infinite autoscaling handling
