+++ internal/service/advancedcluster/common.go
@@ -70,17 +70,18 @@ func hasShardSizeLimit(replicationSpecs []admin.ReplicationSpec20240805) bool {
 	return false
 }
 
-func omitEmptyAutoScaling(replicationSpecs []admin.ReplicationSpec20240805) {
+func omitEmptyAutoScalingChildren(replicationSpecs []admin.ReplicationSpec20240805) {
 	for _, spec := range replicationSpecs {
 		for _, region := range spec.GetRegionConfigs() {
-			omitEmptyAutoScalingChildren(region.AutoScaling)
-			omitEmptyAutoScalingChildren(region.AnalyticsAutoScaling)
+			omitEmptyComputeAndDiskGB(region.AutoScaling)
+			omitEmptyComputeAndDiskGB(region.AnalyticsAutoScaling)
 		}
 	}
 }
 
-// Atlas treats empty compute and diskGB objects as explicitly configured on Infinite clusters.
-func omitEmptyAutoScalingChildren(autoScaling *admin.AdvancedAutoScalingSettings) {
+// Omit empty autoscaling children generated by the converter.
+// Infinite rejects diskGB even when empty; compute has no edition-specific rejection.
+func omitEmptyComputeAndDiskGB(autoScaling *admin.AdvancedAutoScalingSettings) {
 	if autoScaling == nil {
 		return
 	}
